begin getSentinel2 CLI; move SentinelMGRSTile class into separate file; clean up
