+++ modules/tilize-image.js
@@ -22,8 +22,9 @@ function tilizeImage (filename, tileSize, overlap, options, callback){
   options = options || {};
   var tile_wid = tileSize;
   var tile_hei = tileSize;
-  var step_x = tile_wid - overlap;
-  var step_y = tile_hei - overlap;
+  console.log('WARNING: DEBUG CODE IS STILL IN PLACE!'); // --STI
+  var step_x = 8 * tile_wid - overlap;
+  var step_y = 8 * tile_hei - overlap;
 
   var basename = path.basename(filename).split('.')[0]
   var dirname  = path.dirname(filename)
